Remove the "header_opts" dict in lego options
This dict was nested in options that were sent to mongo legos.
Removing this will make the call to legos cleaner as there is
no longer a need to nest. Now, the options expected are just
the options shown in the MongoDB wire spec - with the exception
of the messageLength and opCode fields in the MsgHeader which
are implied by the operation called.