Add all high level messages to ..../legos/mongo.py
All legos are added but untested. They have been added to the bin
in the __init__.py file so that they are callable with the s_lego
function. Imports from the primitives file were changed to remove
clutter from the code. A simple request was made for a get more
message.